Move PatchLightmapColorAssert() to ReachToolPatcher utility class
